+++ homework08/src/main/java/ru/otus/spring/homework08/service/BookServiceImpl.java
@@ -7,7 +7,6 @@ import ru.otus.spring.homework08.domain.Author;
 import ru.otus.spring.homework08.domain.Book;
 import ru.otus.spring.homework08.domain.Genre;
 import ru.otus.spring.homework08.exceptions.IllegalTextException;
-import ru.otus.spring.homework08.exceptions.NoDeleteBookException;
 import ru.otus.spring.homework08.storage.BookDao;
 
 import java.util.List;
@@ -19,6 +18,7 @@ public class BookServiceImpl implements BookService {
 
     private final BookDao bookDao;
     private final IOService ioService;
+    private final CustomService customService;
 
     @Override
     public String insert() {
@@ -55,7 +55,7 @@ public class BookServiceImpl implements BookService {
     public void showByName() {
         ioService.output("Показать книгу с названием");
         String bookname = ioService.input();
-        Book temp = bookDao.getByBookName(bookname);
+        Book temp = bookDao.getByName(bookname);
         ioService.output(temp.toString());
     }
 
@@ -63,10 +63,10 @@ public class BookServiceImpl implements BookService {
     public String deleteById() {
         ioService.output("Удалить книгу с id");
         String id = ioService.input();
-        try {
+        if (customService.checkAvailableComments(id)) {
             bookDao.deleteById(id);
             id = String.format("Книга с id %s удалена из БД", id);
-        } catch (NoDeleteBookException e) {
+        } else {
             id = String.format("Книга с id %s не была удалена", id);
         }
         return id;
@@ -76,15 +76,45 @@ public class BookServiceImpl implements BookService {
     public String deleteByName() {
         ioService.output("Удалить книгу с названием");
         String bookname = ioService.input();
-        Book temp = bookDao.getByBookName(bookname);
+        Book temp = bookDao.getByName(bookname);
         String id = temp.getId();
-        try {
+        if (customService.checkAvailableComments(id)) {
             bookDao.deleteById(id);
             bookname = String.format("Книга с названием %s удалена из БД", bookname);
-        } catch (NoDeleteBookException e) {
+        } else {
             bookname = String.format("Книга %s не была удалена", bookname);
         }
         return bookname;
+    }
+
+    @Override
+    public String deleteGenre() {
+        ioService.output("Удалить жанр с названием (жанр будет удалён из всех книг)");
+        String genreName = ioService.input();
+        List<Book> list = bookDao.getByGenreName(genreName);
+        if (list.size() != 0) {
+            for (Book temp : list) {
+                temp.setGenre(null);
+                bookDao.save(temp);
+            }
+            genreName = String.format("Жанр %s удалён из книг", genreName);
+        } else {
+            genreName = String.format("Жанр %s не существует в списке книг", genreName);
+        }
+        return genreName;
+    }
+
+    @Override
+    public String updateGenre() {
+        ioService.output("Для какой книги обновить/удалить жанр");
+        String bookname = ioService.input();
+        Book temp = customService.checkBook(bookname);
+        ioService.output("Какой жанр добавить (для удаления просто нажать enter)");
+        String genrename = ioService.input();
+        Genre tempGenre = genrename.equals("") ? null : new Genre(genrename);
+        temp.setGenre(tempGenre);
+        bookDao.save(temp);
+        return String.format("Жанр для книги %s обновлён", bookname);
     }
 
     @Override
@@ -101,11 +131,6 @@ public class BookServiceImpl implements BookService {
             mas[i] = mas[i].trim();
         }
         return mas;
-    }
-
-    @Override
-    public Book getBook(String bookname) {
-        return bookDao.getByBookName(bookname);
     }
 
 }
